--- main/src/main/java/org/sopt/makers/crew/main/external/notification/dto/PushNotificationResponseDto.java
@@ -1,20 +1,0 @@
-package org.sopt.makers.crew.main.external.notification.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class PushNotificationResponseDto {
-
-  @JsonProperty("status")
-  private Integer status;
-
-  @JsonProperty("success")
-  private Boolean success;
-
-  @JsonProperty("message")
-  private String message;
-
-}
