+++ main/src/main/java/org/sopt/makers/crew/main/external/notification/dto/response/PushNotificationResponseDto.java
@@ -1,0 +1,21 @@
+package org.sopt.makers.crew.main.external.notification.dto.response;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class PushNotificationResponseDto {
+
+	@JsonProperty("status")
+	private Integer status;
+
+	@JsonProperty("success")
+	private Boolean success;
+
+	@JsonProperty("message")
+	private String message;
+
+}
